streamlit/requirements.txt
streamlit==1.36.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pillow==10.4.0 